+++ src/App.jsx
@@ -1,8 +1,8 @@
 
 import { useState } from 'react';
 import './App.css';
-import TodoList from './TodoList.jsx';
-import TodoForm from './TodoForm.jsx';
+import TodoList from './features/TodoList/TodoList.jsx';
+import TodoForm from './features/TodoForm.jsx';
 function App() {
 const [todoList, setTodoList]= useState([]);
   function addTodo(title){
